sample-apps/web-app/requirements.txt
flask==3.0.0
pyopenssl==24.0.0
werkzeug==3.0.1
